+++ docs/.vuepress/nav/version5x.js
@@ -49,8 +49,8 @@ module.exports.sidebar = function(){
             children: [
                 'container/',
                 'container/install',
-                'container/reg-as-app',
-                'container/destroy',
+                'container/service-container',
+                'container/list-resolver',
             ]
         },
         {
@@ -137,6 +137,7 @@ module.exports.sidebar = function(){
                                 'http/clients/methods/attachments',
                                 'http/clients/methods/cookies',
                                 'http/clients/methods/expectations',
+                                'http/clients/methods/middleware',
                                 'http/clients/methods/conditions',
                                 'http/clients/methods/criteria',
                                 'http/clients/methods/redirects',