+++ packages/Http/Clients/src/Drivers/BaseClient.php
@@ -3,7 +3,10 @@
 namespace Aedart\Http\Clients\Drivers;
 
 use Aedart\Contracts\Http\Clients\Client;
+use Aedart\Contracts\Http\Clients\Middleware;
 use Aedart\Contracts\Http\Clients\Requests\Builder;
+use Aedart\Contracts\Http\Clients\Requests\Handler;
+use Aedart\Http\Clients\Requests\Handlers\QueueHandler;
 use Aedart\Support\Helpers\Container\ContainerTrait;
 use Illuminate\Contracts\Container\Container;
 
@@ -81,4 +84,34 @@ abstract class BaseClient implements Client
     /*****************************************************************
      * Internals
      ****************************************************************/
+
+    /**
+     * Creates a new handler that is able to process middleware for
+     * outgoing request and incoming response.
+     *
+     * @see Handler
+     * @see Middleware
+     *
+     * @param  Handler  $fallbackHandler The fallback handler
+     * @param  Middleware[] $middleware [optional] Middleware to be executed
+     *
+     * @return Handler
+     */
+    protected function makeMiddlewareHandler(Handler $fallbackHandler, array $middleware = []): Handler
+    {
+        return (new QueueHandler($fallbackHandler))
+            ->addMultiple($middleware);
+    }
+
+    /**
+     * Extracts middleware from driver options
+     *
+     * @param  array  $options Driver options
+     *
+     * @return Middleware[] Empty if none available in options
+     */
+    protected function middlewareFromOptions(array $options): array
+    {
+        return $options['middleware'] ?? [];
+    }
 }