+++ packages/Http/Clients/src/Drivers/DefaultHttpClient.php
@@ -2,8 +2,11 @@
 
 namespace Aedart\Http\Clients\Drivers;
 
+use Aedart\Contracts\Http\Clients\Middleware;
 use Aedart\Contracts\Http\Clients\Requests\Builder;
+use Aedart\Contracts\Http\Clients\Requests\Handler;
 use Aedart\Contracts\Http\Clients\Requests\HasDriverOptions;
+use Aedart\Http\Clients\Requests\Builders\Guzzle\Handlers\SendRequestHandler;
 use Aedart\Http\Clients\Requests\Builders\GuzzleRequestBuilder;
 use GuzzleHttp\Client as GuzzleClient;
 use GuzzleHttp\RequestOptions;
@@ -78,8 +81,11 @@ class DefaultHttpClient extends BaseClient
             $options = $request->getDriverOptions();
         }
 
-        // Perform actual request
-        return $this->driver()->send($request, $options);
+        // Prepare middleware handling and handle outgoing
+        // request and incoming response
+        return $this
+            ->prepareMiddlewareHandler($options)
+            ->handle($request);
     }
 
     /**
@@ -98,5 +104,36 @@ class DefaultHttpClient extends BaseClient
     public function makeBuilder(): Builder
     {
         return new GuzzleRequestBuilder($this, $this->getClientOptions());
+    }
+
+    /*****************************************************************
+     * Internals
+     ****************************************************************/
+
+    /**
+     * Prepares the middleware handler
+     *
+     * @param  array  $options Driver options
+     *
+     * @return Handler
+     */
+    protected function prepareMiddlewareHandler(array $options): Handler
+    {
+        return $this->makeMiddlewareHandler(
+            $this->makeFallbackHandler($options),
+            $this->middlewareFromOptions($options)
+        );
+    }
+
+    /**
+     * Make the fallback handler for when processing middleware.
+     *
+     * @param  array  $options  [optional] Request options
+     *
+     * @return Handler Responsible for sending request and returning received response
+     */
+    protected function makeFallbackHandler(array $options = []): Handler
+    {
+        return new SendRequestHandler($this->driver(), $options);
     }
 }